+++ lib/rails/test_unit/sub_test_task.rb
@@ -1,0 +1,4 @@
+# Monkey patch: waiting for Draper to be full comptable with Rails 5
+# Related issue: https://github.com/drapergem/draper/issues/681
+class Rails::SubTestTask < Rake::TestTask
+end
