+++ spec/decorators/question_decorator_spec.rb
@@ -1,0 +1,7 @@
+require 'rails_helper'
+
+RSpec.describe QuestionDecorator do
+  describe '#last_answers' do
+      
+  end
+end
